Update patch set 3
Patch Set 3:

Eric, my previous fix for Offloaded/VolumeShaper was to ensure advancement when muted.

This CL changes the timing of volume processing.  It is a bit risky to do this change because volume should be able to change while the track is active.

Patch-set: 3
Reviewer: Gerrit User 1040213 <1040213@85c56323-6fa9-3386-8a01-6480fb634889>