Check the address for same device routing skip
When the address of the device is different from that of the Patch sink,
the routing should not be skipped even though the device type is same.
In Automotive Android, where the different devices have the same type
AUDIO_DEVICE_OUT_BUS, routing changes would be skipped based on the device
type check.
Scenario : Whan a call is started, all audio patches are updated with
the same sink, but at the end of the call, when rerouting to the respective
sinks, the setOutputDevice skips the routing update based on the above check.
To prevent this, along with the check for device type, the address should
also be compared.

Test: Start a call, make sure the devices are routed correctly.
After call, play media, navigation usecases and verify the device routing.

Change-Id: I6000b49d4ef17ce26cefd76a2b3150164e04fc00
Signed-off-by: anoobk <anoob.anto.kodankandath@intel.com>