Update patch set 3
Patch Set 3:

The current version of CL may not work if the buffer queue lives in the same process as ACodec. Even though that doesn't seem to be the case in the settings tested, I cannot say the same in general.

This problem seems to have stemmed from the ambiguity in the semantics of native_window_set_sideband_stream. I think we should fix it properly.

One way I recommend is to specify in the comment block above native_window_set_sideband_stream that the native handle will always be cloned; then we make sure that this new requirement is satisfied by always cloning the handle in Surface::dispatchSetSidebandStream instead of just taking a reference.

Patch-set: 3
Reviewer: Gerrit User 1105338 <1105338@85c56323-6fa9-3386-8a01-6480fb634889>