Update patch set 2
Patch Set 2:

I still don't understand the purpose of this CL. What "missing symbols" are you talking about?
Currently, e.g. libamrextractor.so is not linked with libutils, either statically or dynamically, and does not use any symbols from libutils. So what's the point of statically linking libamrextractor.so with libutils, when it's not going to use any symbols from it?

Patch-set: 2
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>