Update patch set 3
Patch Set 3:

also, given that this is about dependencies from libstagefright_foundation (even if they're ultimately unused/unneeded/discarded), my earlier comment about requiring "do not merge" was wrong, and it should be removed again.

I still think the toolchain shouldn't require symbols from code that's not going to be in the final binary, but as long as this change doesn't increase the size of that final binary, I suppose we can live with it.

Patch-set: 3
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>