Camera: Wait until the current request id appears in the results
Camera preview stop only waits until the
current capture request is sent to Hal and
then immediately flushes the device.
Depending on timing and the pending request
queue size within CameraHal the current camera
client parameters may get flushed before
they are able to reach the end of the
processing pipeline.
Avoid this by waiting for the current request
id to appear in the incoming results.

Bug: 139330668
Test: Camera CTS, manual using application

Change-Id: I18de85557e8e76bfbe6d69b1eb2a5c36ed64e803