Update patch set 2
Patch Set 2:

sorry for this, but I am not sure how we want to handle this. Listing /system/{LIB} for spgal to old ld config will work, but that is not a ideal solution. So I come up with this, but not sure how we could handle the vndk flags in bp. Any ideas?

Patch-set: 2
Reviewer: Gerrit User 1319385 <1319385@85c56323-6fa9-3386-8a01-6480fb634889>
Work-in-progress: false