Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The bug says it's about moving static libcrypto users to libcrypto_static, and indeed that's what some of the CLs linked from the bug do.
> This CL however moves the mpeg2 extractor from statically linked libcrypto to dynamically linked libcrypto. Is that intentional?
> 
> Since the mpeg2 extractor is part of the media apex, will this pull libcrypto.so (and whatever dependencies it may have) into the apex as well? Is that OK/expected? (it seems to me that if there are concerns with including a statically linked libcrypto, those would also apply to a dynamically linked libcrypto bundled in the apex)

This may need an additional change to add libcrypto to the apex.  An extra copy of dynamically linked libcrypto is not a problem for FIPS, since it is built the same as the normal libcrypto.so.  We are trying to remove any uses of static libcrypto from the core platform.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>