C2SoftDec: Use stride returned by graphic view in avc, hevc and mpeg2
Instead of assuming stride to be a multiple of 64, use the value
returned by graphic view.
These codecs require chroma stride to be half of luma, hence the luma stride
should be a multiple of 32. Hence request for stride that is multiple of
32, then use returned stride.
This is needed where allcoated buffer has a stride that is multiple of
128.

Test: atest android.media.cts.DecoderTest
Test: Test decoding clips with dimensions that are not multiples of 16

Bug: 144190181
Change-Id: I08912396e495326fca787e8ae0b47256505210ca