Update patch set 2
Patch Set 2:

(7 comments)

the return vs continue in C2DecoderTest.cpp is the one that should be fixed before we insert this. The other comments can be done now, or can be handled with a subsequent CL in a subsequent bug (if done this way, create the bug and reference it in a response to the appropriate comments in this code review).

Patch-set: 2
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>