stagefright: make VideoFrameScheduler for near constant vsync remainders
The edge-remainder simplification assumes a somewhat scattered vsync
offset remainder over the next few frames, which is not the case
if the video period is a near multiple of the vsync period. This
can cause an error in selecting a proper correction-limit. For this
specific case the vsync offset is expected to span the whole vsync
range. Simply use 1 frame in this case as that gives the best
prediction of the average offset.

Bug: 140789526
Test: semi-automated test evaluating original set of video/display-rate
	scenarios now including 61Hz refresh rate. To be submitted.
Change-Id: I5897e1c070c3d95112992997b1307867a415f882