Check if PatchTrack has enough data to write once
On AudioPlaybackCapture, underrun can occurs
in secondary PlaybackThread during playback.

When primary PlaybackThread is blocked
by any reason during playback,
interceptBuffer is also blocked and
data come to be not provided to PatchRecord buffer.
But, secondary PlaybackThread is running,
so data in PatchTrack exhausts.

In this time, mFillingUpStatus of PatchTrack keeps FS_ACTIVE,
so write process is continued even though framesReady()
is not filled to threshold(mFrameCountToBeReady).
So, underrun happens repeatedly.

To avoid this issue,
check if track has enough data in releaseBuffer().
And, if no enough data to write once,
change mFillingUpStatus to FS_FILLING,
and wait for filling buffer.

Bug: 144009535
Change-Id: Ifcbaea05228064284dfaa9b629c8860b1fbe977d