WIP rework media tests for for atest and TEST_MAPPING
Rework test code so that it fits into the atest and TEST_MAPPING
frameworks. This includes automated pushing of the test content to the
device along with the test itself. This changes what were previously
warnings into assert() calls that signal an error.

There may still be some debugging diagnostics and comments in these
files, so some cleanup could be needed (along with this commit message
once done).

Bug: 140051680
Test: atest
Change-Id: I0bd6c386656a622b1362e4581dd0de64649f94d9