Update patch set 11
Patch Set 11:

Ray,
All the comments have been responded to for these three CLs (this and two CLs that this depends on)
Can you please take another look and submit these?
We will submit a follow up CL for for converting the WARN prints to ASSERTS() for tests that are already merged.

Patch-set: 11
Reviewer: Gerrit User 1047003 <1047003@85c56323-6fa9-3386-8a01-6480fb634889>