[AudioPolicy][EngineConfigurable] Prevent regression on audio-base.h change
This CL prevents from regression each time audio-base.h is changed on
enumeration coupled with Parameter Framework structure.

The Structure type file is generated from the audio-base.h. If misaligned,
Settings would not be generated, it won't compile.

This CL also fixes a typo on name as names are now aligned.

Bug:
Test: build

Change-Id: I8fd15a8d2f92908dceb2baf7ca8d35f4a69f7904
Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>