Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (2 comments)
> 
> question: is the android device enum to pwf device mask conversion relying on the fact that only one bit is used for a given device in Android enum? 
> If yes we need to change that given the new device type representation.

Yes, definitely, we do use the bitfield for inclusive criterion. As for now, the deviceTypesToBitMask utility does the job for us. But I see that there is a plan to remove it completely.

We developped in the past an inclusive criterion not based on bitmask. This was never merged :-(
We also had a plan to move to 64bit criterion, but it is just delaying the failure point.

So, might be **urgent** to do a revival of the first development...

Patch-set: 3