Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

would be more excited if we knew which change (hidl code or toolchange) 'fixed' it,
but that's more 'wish' than 'need'. But I'm  not sure that the effort to narrow it down is worth it.

would like to retain some sort of pointer in the event that we take this code back to compile in 
a Q-based environment (as may be required as part of the mainline world).

Patch-set: 1
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1