Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1: Code-Review+1
> 
> (1 comment)
> 
> would be more excited if we knew which change (hidl code or toolchange) 'fixed' it,
> but that's more 'wish' than 'need'. But I'm  not sure that the effort to narrow it down is worth it.
> 
This doesn't seem to be fixed from a toolchain change because I could build this change in the Q branch with the toolchain from Feb 2019, when coverage was originally disabled here.

I agree that it'd be good to know how this got fixed.  I'm unaware of any tooling that could help with this though.

Patch-set: 2
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>