Workaround: VolumeSource without stream type fixup
VolumeSources without legacy stream type assigned need to choose ony stream
(default aka MUSIC) in order to set the SW gain within AudioFlinger.

If another source is currently active on the current SwOutput with Music stream
type, the gain will be overwritten.

This CL is a temporary workaround: it prevents to set the gain of an unactive
volume source on a given SwOutput if another source is active.

The clean fix would be to migrate AudioFlinger to manage gain per volume Source (or
attributes).

Test: build & play audio
Bug:
Change-Id: Iabaeb2d4fe5fc384f6eabd5058aceba6303fcdf6
Signed-off-by: François Gaffie <francois.gaffie@renault.com>