Update patch set 5
Patch Set 5:

I was thinking of the actual problem:
It should not be permitted to have the same stream in two different volume groups right?
So this thing of defaulting volume groups without streams to MUSIC does not work
And why do we get volume groups without streams in the first place?
Why do we end up with duplicated volume groups for patch en rerouting?
Something must be broken in the volume group init code in default engine.

Patch-set: 5