Update patch set 1
Patch Set 1:

this changes semantics -- this looks to be a test to decide if we should skip an iteration as we
walk through a bit mask; the old semantics seem to me to never fire, so we'd never skip an iteration.
I don't know if that was just a loss of an optimization (did some useless work) or if there are
some bad consequences from doing the guts of that loop iteration.

first blush is that the old semantics were incorrect -- but want some more folks to look at this
to be sure.

Patch-set: 1
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>