Update patch set 4
Patch Set 4:

(6 comments)

thanks for the excerpt from the standard -- it explains what is allowed to be out there at the end of the tx3g box.

i think the use of 'remaining' within the ftab box needs to be changed to use the size field that is at the front of the ftab box.  that's one i do want to see changed.

Patch-set: 4
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>