Update patch set 2
Patch Set 2: Code-Review-2

I believe this CL is unnecessary.
The position should be a 32-bit frame counter that wraps at 0x7FFFFFFF.
It is not the position of the DSP in the buffer.
It is the position within the stream.

For more details, see: https://buganizer.corp.google.com/issues/150105878#comment19

Patch-set: 2
Reviewer: Gerrit User 1060750 <1060750@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2