Update patch set 1
Patch Set 1:

@Eric: I came with the conclusion that volume could not be managed on Hw Audio Bridge.
Starting the dev on HwAudioOutputDescriptor was quite huge, and I decide to make the maximum reuse of the classes you introduced to manage activity for Audio Source.

Here is a proposal to solve this issue reusing the Client of audio source, forcing the call also to create a source client (whatever the way it is routed, legacy setOutputDevice / audiopatch SW / HW).
It allows even not to make the voice call a special case in setVolumeIndexForAttributes.

This proposal is using the fact that all sink device have a mixPort dedicated in order to be declared as reachable and hence available.
The Client Source Desc is added on this mixPort.

It allows also to use the priority between strategy to set the volume, when the volume is HW, to prevent overwritting the bus gain if more than 2 use cases are addressing the same sink device.
(It can be a Device to Sink Device and a mixPort to this same sink Device)

I tested on Emulator on all type of AudioPatch, emulating Voice call in all way of routing + re run 
the functional test (that are not merged...).

Patch-set: 1
Reviewer: Gerrit User 1226222 <1226222@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1598908 <1598908@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1597543 <1597543@85c56323-6fa9-3386-8a01-6480fb634889>