Update patch set 1
Patch Set 1:

(2 comments)

it feels like this is about to involve changes to AndroidTest.xml and the environment setup code.

I see a WriterTest.cpp:#define OUTPUT_FILE_NAME "/data/local/tmp/writer.out"
but i don't see anywhere that we remove it after a test is done; that's something that
should be addressed.
If it's desirable as something to figure out why things are broken, then add an invocation
flag that says to keep it -- but the default should be to delete it.

Patch-set: 1
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>