Update patch set 9
Patch Set 9:

(if we're depending on the AndroidTest.xml file to clean up the downloaded files and
that these live in that directory.... i'd rather not do that.

that only works for the atest invocations; it won't work for the manual invocation.

last week we added a '-C' option to one of the tests so that it would clean up afterwards, 
maybe we need to a similar construct here (and it should default to cleaning up these intermediate files).

Patch-set: 9
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>