Update patch set 3
Patch Set 3:

Kris,
This is now ready to be merged.
Can you please review and submit this?
Once this is submitted, I will trigger presubmit tests on mp3 and aac that depend on this CL.
Without a +2 on this, I can't trigger presubmit runs on those CLs.

Patch-set: 3
Reviewer: Gerrit User 1047003 <1047003@85c56323-6fa9-3386-8a01-6480fb634889>