Update patch set 2
Patch Set 2:

internal/master already does this.
the CL commit message is:
>> Link with actual headers instead of symlinks
>> 
>> and remove the symlinks.
>> 
>> Bug: 143245419
>> Test: build
>> Change-Id: I70e450d9db0f65048f86391c4d3f341659cd213f

should have a merged-in line on this, so that we know we don't need to merge into master.

it merged long enough ago, that you should see this in the partner r-fs branches

Patch-set: 2
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>