[Bugfix]add onBufferDetached callback to release the <sp>GraphicBuffer reference in cameraservice and add bufferqueue3.0 support for libstagefright_bqhelper
if call acquireBuffer to acquire GraphicBuffer from BufferQueue1, and then call ImageWriter.queueInputImage to do prevOwner.detachImage, and call attachAndQueueInputImage
to do attach and queue the GraphicBuffer to BufferQueue2,
we can only detach the BufferQueueCore's <sp>GraphicBuffer reference from BufferQueue1,
but can not release the <sp>GraphicBuffer reference that mSlots[slot].buffer in Surface from BufferQueue1 until Surface is destroyed.

In this case, the GraphicBuffer not be released causes memory redundant use.
So we add onBufferDetached callback method interface support in IProducerListener
and add releaseSlot method in Surface to release the <sp>GraphicBuffer reference.

Test: the test results:We can release 600M memory at xiaomi Camera, this issue can also test in two BufferQueues scene,as mentioned above.

Signed-off-by: Wangrui <wangrui26@xiaomi.com>
Change-Id: I601a6d97eb32c9ca732c95ed39236a0ed1e73cfc