AudioPolicy: manager: force gain on startSource when using HW Gain
When a source is started, the volume is evaluated for the given
volume source on the SwOutput. If the gain index (cached) has not changed
for the given source, the index is not applied again.
However, in case of volume source, to manage the conflicts, a guard
on the activitiy of this source based on the clients prevent
to apply it on AudioHAL.
If the volume is changed before source is started, it will never be
sync'ed in AudioHAL until any change of index happens.

This CL forces gain application on SwOutput only if SwOutput is using
HW gain (aka routed on device on which gain is managed by HAL, not
by SW in AudioFlinger)

Test: build & basic audio tests.

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ief1d5314b7970de78298750550320e329283ee1a