Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> this doesn't quite feel right to me.
> I remember a lot of "remove binder dependency" work last year was really more of a "convert binder to ndk binder".
> and if wedo decide to remove these.... vndk doesn't feel like the right condition.
> 
> other notes:
> -- the bug this references was closed a few months ago
> -- i believe this is for fuzzing purposes, which is largely tied with host_support:true
> 
> adding some other people to look at this.

Ray,
This change is cherry-picked from r-fs-release and hence the bug number corresponds to an issue that was closed few months ago.
Should we file a new bug?
Yes, this change is to add a host_supported:true for libstagefright_foundation that is needed to enable host support for extractors and then eventually for extractor fuzzers.

As Android.bp has changed for libstagefright_foundation in the internal branches, bringing this change to aosp/master helps resolve the merge conflict

Patch-set: 2
Reviewer: Gerrit User 1047003 <1047003@85c56323-6fa9-3386-8a01-6480fb634889>