AudioPolicy: remove devices while Sw bridge established
When devices are disconnected while involved in a SwBridge,
the connectAudioSource function may be called while the cache
of the engine has not yet been updated with availability of devices.

This CL forces to retrieve sinkDevices from engine without using
the cache.

It also ensure in case of closeOuput called before a Sw bridge patch
is released that the patch will be released first.
It prevents Audio HAL to receive releaseAudioPatch involving
an output that does not exist anymore.

Test: AudioPolicyEmulatorTests --gtest_filter=AudioPatchTest.*

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I2663d10f08de3191de4d95bc81749ede9b8a360f