removed the exception handling to skip the frame which data size is 0
removed the exception handling that stop finding frame if frame data size is 0 on parsing id3 tag v2_3/v2_4.
When finding frame stopped, some frames following the frame (data size is 0 and whole size is 10-header size) cannot be extracted.
If the frame is corrupted, it can be checked through other exception handling.
In the market, there are a lot of mp3 files with some frame data size is 0 on id3tag.
If skip this frame, the entire frames on id3tag can be extracted normally.

Bug: 148559957
Test: manual - check the details(meta data) of mp3 file with the frame data size is 0 on id3tag.

Signed-off-by: Jiyoung <jiyoung0720.bang@samsung.com>
Change-Id: I9c43fc7ef899b9f447538eb8c9631a68876d89e2