Enable ION_FLAG_CACHED flag for allocated ION bufers
Enable the ION_FLAG_CACHED flag for allocated ION bufers. It is required
because the read/write operations from/to ION buffers are executed by
CPU from different threads, and the enabling of this flag allows
to keep a cache coherency and avoids a data corruption in the buffers.
The patch applying leads to all tests of the VtsHalMediaC2V1_0Host
module are passed successfully.

Test: run vts -m VtsHalMediaC2V1_0Host
Change-Id: Ic855884c19d3c5ffeffe588f58a245aa785bfb83
Signed-off-by: Oleksandr Bobro <oleksandr.bobro@globallogic.com>