Update patch set 2
Patch Set 2: Presubmit-Ready+1

I split the change for codecs_g711dec(aosp/1322145) because it needs to be set, which was missing in the original change. Note that this is a cherry-pick and it would be automerge-skipped in the internal branches.

Patch-set: 2
Reviewer: Gerrit User 1415877 <1415877@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1