Register timestamp discontinuity on pause
During AudioTrack pause(), the HAL presentation position is not
requested by AudioFlinger until AudioTrack resume() occurs. When this
happens, the elapsed time since pause may be quite long causing the
timestamp verifier to estimate a jitter correction that has a
correspondingly large error. This causes the corrected timestamp to
jump.

After resume, in order to prevent retrograde timestamps in
AudioFlinger, subsequent timestamps will be discarded until the values
surpasses the erroneous jitter corrected timestamp described above.
This results in A/V sync drift.

In order to avoid such jitter correction to the timestamps, register a
timestamp discontinuity on pause so that the initial timestamps after
resume do not undergo jitter correction.

Change-Id: I8a4ecbcc216eb1f8fdf4be80416a825daee434cb