Update patch set 3
Patch Set 3:

Order shall not matters...
Do you means AudioAttributes#toVolumeStreamType is reporting the wrong stream type?
I remember this kind of bug:

https://android-review.googlesource.com/c/platform/frameworks/base/+/1243111

Normally it is merged, so if this issue is present and this CL merged, I will analyse it. 

However, in the meantime, let s invert it as a temporary WA.

Patch-set: 3