Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3:
> 
> Order shall not matters...
> Do you means AudioAttributes#toVolumeStreamType is reporting the wrong stream type?
> I remember this kind of bug:
> 
> https://android-review.googlesource.com/c/platform/frameworks/base/+/1243111
> 
> Normally it is merged, so if this issue is present and this CL merged, I will analyse it. 
> 
> However, in the meantime, let s invert it as a temporary WA.

Sorry, I understood. AUDIO_USAGE_ASSISTANT is also a rule of matching for MUSIC whatever the content type is.
ASSISTANT stream matching is stronger, so as filtering on match on usage and match on content type.
So order matters in this case. 
So, not a WA, real fix.

Patch-set: 3
Reviewer: Gerrit User 1491204 <1491204@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1