Support direct PCM track latency reporting
For MSD -> primary HAL software patches, it is useful
to know the primary HAL direct output thread track
latency compared with the total software patch latency,
in order to determine the contribution of primary HAL
latency to the overall MSD SW patch latency. Allow this
to be reported for MSD patch playback threads that
support it (PCM non HW A/V sync).

Test: MSD patch with PCM non HW A/V sync output
      audioflinger dumpsys track attached to direct
      output thread shows track latency.
Change-Id: I9b271de446993dd21a0947e4d68b7a6ac23fb30c