Update patch set 7
Patch Set 7:

it feels like the routines that this fuzzes are all stubs that invoke across a binder to
various other services.  Is our goal here to test the code this calls directly or to test
the calls over on the other side of those binder calls?

I'm not against either answer, I just want to be clear about which entity (or entities) we think
we're testing....

Patch-set: 7