Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> it feels like the routines that this fuzzes are all stubs that invoke across a binder to
> various other services.  Is our goal here to test the code this calls directly or to test
> the calls over on the other side of those binder calls?
> 
> I'm not against either answer, I just want to be clear about which entity (or entities) we think
> we're testing....

The goal is to fuzz this library and the components within it, not to test binder. However, many of the features of this library are essentially just interfaces for calling out to binder, so that's definitely going to be part of this. So to answer your question, yes and no. I am not intentionally targeting binder (there's another fuzzer in the works for that currently), however I am trying to hit existing code paths/functionality contained within this library, which does often call out to services through binder.

Patch-set: 7