Update patch set 5
Patch Set 5:

(2 comments)

the thing I noticed when looking at one of the new clauses I asked you do add [and thanks for adding that] is that there are a number of places where we haven't reset some variables before our next test -- putting us in a position where we can't say that a variable has the right answer because the most recent test put it there..... or if the most recent test failed to modify it and it still has the value from the previous test.

So I think the next step is to walk through here and make sure that we're always resetting properly
before each test.  my comments around line 376 show one example of how we need to reset.

Patch-set: 5