Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 3:
> > 
> > Dean, I ran `audiopolicy_tests` unit tests to smoke test your change, and AudioPolicyManagerTestMsd#PatchCreationOnSetForceUse started failing reporting that an extra patch is being created. I haven't yet time to look into the reasons behind this.
> 
> Hi Mikhail - I'm not seeing this error when running these tests ($atest audiopolicy_tests -c -v) on hikey arm or cuttlefish x86. Should I use a different method to invoke the test failure?

Thanks for checking! I'm seeing this failure when trying this change on internal master so it's possible that this failure happens due to interaction with some unpublished change. I will dig deeper. The device you run the test on shouldn't matter because the test uses its own APM configuration.

Patch-set: 4