WIP: Codec2: C2Buffer: Add logic to select C2DmaBufAllocator when appropriate
This patch is not great, and suggestions for improvement are
welcome.

This trys to add logic to the C2Buffer code to switch between
ION and DMABUFHEAP allocators as appropriate. Unfortunately the
logic switching bewteen BLOB and ION code is already complex so
adding an extra case here isn't pretty.

Change-Id: I2dee23bd2cf5659a0bc140eb42ac6d93bd77d460
Signed-off-by: John Stultz <john.stultz@linaro.org>