WIP: codec2: vndk/C2Buffer: Rework CreateLinearBlock to use fetchAllocator()
The current code in _C2BlockFactory::CreateLinearBlock()
duplicates logic in C2Store code and creates extra complexity to
extend it for the future DMABUF Heaps case.

So clean it up by getting rid of the lambda and using
fetchAllocator() instead.

NOTE: I managed to re-trigger the crash I was seeing with this
so clearly something is wrong with it.

Close review would be appreciated!

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I7c3cc40fe82bd0b7646ff62c3456b158682e04e0