WIP: codec2: vndk/C2Buffer: Rework CreateLinearBlock to use fetchAllocator()
The current code in _C2BlockFactory::CreateLinearBlock()
duplicates logic in C2Store code and creates extra complexity to
extend it for the future DMABUF Heaps case.

So clean it up by getting rid of the lambda and using
fetchAllocator() instead.

NOTE: I do see occasional crashes with this patch faulting on:
  frameworks/av/media/codec2/vndk/C2Buffer.cpp:439
usually as a video finishes playing. I need to dig more on what
is failing (maybe I still do need the lambda?)

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I7c3cc40fe82bd0b7646ff62c3456b158682e04e0