audiopolicy: Optional fallback on default strategy/group
When attributes does not match a group or strategy, it fallbacks on
respectively default group or strategy.

This CL allows to control whether fallback shall be done or not.
It allows for example to detect orphans attributes and prevent overwrite
volume for example.

Bug: 179759299
Test: make

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ieeeab48ca3ec2b031b7a99e2e7ab26d2e4d70535
Merged-In: Ieeeab48ca3ec2b031b7a99e2e7ab26d2e4d70535