[AudioFlinger] Control volume using Port ID
This CL migrates the volume management within AudioFlinger
from stream type to port ID.

It gives full power to AudioPolicy to compute the list of port
(so MmapThreads/Tracks) on which volume control is required.

It prevents from overwritting MUSIC stream type which is
the default for volume groups without associated stream type.

Test: build & play audio

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I09bd24c9a03633b5df939544e857be4409b33623