audiopolicy: HW Bridge Sink matching rule
In order to use an HW Bridge, audio policy configuration file
must declare that a given sink may be reacheable from a given source
device.
The declaration is made using the "Tag Name", which is a human readable
name given to a device / mix Port.

This CL adds an equality operator based only on this tag name, which
identify uniquely the device port.

This CL manages also dynamic device cases, the tagName shall be assigned
to all dynamically created device.

Bug:

Test: adb shell ./data/AudioPolicyEmulatorTests/AudioPolicyEmulatorTests --gtest_filter=*VoiceCallBridgingTest*

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I0fc6435dcc5f40634be5392dd92dcd2f377809c2