audio: fix direct output resource be ocupied too long
While flush is follow after pause, mState might be overwrited and not
processed. In this case, direct output will retry due to data underrun.
In the duration of retry, the ActiveCount in AudioPolicy will be ocupied
and user can't create new track in time.
Add isPausePending to make sure pause will be well handled in direct output.

Test: Manual (see bug)
Test: Clarity MediaPlayer Offloaded with seek.
Bug: 175092336
Change-Id: Ief742d2ce44917da24885373a5770e47876bc3a5