av: fix static analyzer complaints
This CL fixes two static analyzer complaints in av code.

> frameworks/av/media/codecs/amrnb/enc/src/cor_h_x2.cpp:278:26: warning:
1st function call argument is an uninitialized value
[clang-analyzer-core.CallAndMessage]

This is a valid complaint, but we force all uninitialized memory to
zero, so it seems harmless to make this explicit in the code.

> frameworks/av/media/extractors/mp4/MPEG4Extractor.cpp:4663:32:
warning: Access to field 'meta' results in a dereference of a null
pointer (loaded from field 'mLastTrack')
[clang-analyzer-core.NullDereference]

when saving things into `tempForwRefVop` and `tempRefSelCode` above,
there's a NULL check, so it's probably best to have down here, too.

Bug: None
Test: TreeHugger
Change-Id: Iec30bfa80af4424afedb68d85f06c7134eaa2e09