Stagefright: WA to use different media_codecs xml
This is a common issue. This patch is to config 4k capability for GSI test
cases.

Workaround to read different media_codecs xml files on various targets
in GSI and non-GSI cases.
This is to avoid using very high configuration line 8K on GSI, where
there are some memory related issues in AOSP.

Bug: 178725280

Test:
run cts-on-gsi -m CtsMediaTestCases
    -t android.media.cts.VideoEncoderTest#testOtherH264FlexNearMaxMax
    -t android.media.cts.VideoEncoderTest#testOtherH264SurfNearMaxMax
    -t android.media.cts.VideoEncoderTest#testOtherH265FlexMaxMax
    -t android.media.cts.VideoEncoderTest#testOtherH265FlexNearMaxMax
    -t android.media.cts.VideoEncoderTest#testOtherH265SurfMaxMax
    -t android.media.cts.VideoEncoderTest#testOtherH265SurfNearMaxMax

Change-Id: Ibb4c75ed910910a5e3da625ddc48fb0452b08e82