[AudioPolicyManager] Voice Downlink Sink device selection
Audio Policy selects the downlink sink device through the primary
output of the primary module.
It uses getNewOutputDevices on this output. If any routing strategy is
active on this output with an higher priority than voice call (it may
happen as order is defined in engine configuration file), downlink may
be temporarilly routed on this sink until the other client stops.

This CL changes the downlink sink device selection by
-checking if a HW bridge is possible between the selected sink device
for voice call strategy and telephony RX device
-falling back on "legacy" behavior if no HW bridge possible.

Test: call with and without music active.

Change-Id: I8b6a6648a72f6f69abdf4d49c205ccbd0efeb082