[AudioPolicyManager] Prevent to fallback on music group for orphan streams
When using a specific engine configuration file with custom product strategies
definition, nothing prevent not to attach a given stream type to a strategy.
In order to prevent fallbacking on music to consider activity of non existing
stream, this CL adds an optional fallback parameter to stream to volume source
conversion APIs.

Test: plays with volumes

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I0817586a5ac3f2a8be26bc9fb21ba1c079d28006