[AudioPolicyManager] Do not fallback on music group for orphan streams
When using a specific engine configuration file with custom product strategies
definition, it is possible to define a strategy without any stream type attached
to it.
In order to prevent fallback on music when considering activity of non existing
stream, this CL adds an optional fallback parameter to stream to volume source
conversion APIs.

Test: plays with volumes

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I0817586a5ac3f2a8be26bc9fb21ba1c079d28006