[AudioPolicyManager] no routing in case of combo UC
If an Output descriptor is used for volume control while an
Audio Source is active and realized through an Hw Bridge (Device
to Device audio patch), this output is declared as active but
is not really routed. Laterly, it will prevent any client to be
effectively routed on this output.

Test: Voice Call through Hw Bridge mixed with another playback
routed on same output managing voice call volume.

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: Ib95b4ca82c0792f3f0f223dc36c2132fac011b88