AudioFlinger: Patch Panel: prevent to remove Plaback patch of a SW Bridge
When a SW Bridge is created, 2 patches are created. The patch involving
the SwOutput and the sink device is attached to the SwOutput mix handle.
When this SwOutput is routed/rerouted, the setOutputDevices will create
also a new Patch on AudioFlinger.
The remove stale audio patch section may destroy this patch.
Thus, when releasing the SwBridge patch, the Playback patch is not found
and releaseAudioPatch is never called on AudioHAL.

This CL prevents to erase the Playback Patch of a Sw Bridge.

Also, when a SwOutput is already routed (aka an output device is attached/
patch is associated), AudioFlinger will destroy this patch once switching
to the SwBridge.
If SwOutput is routed after bridge creation, APM shall release both APM/AF
patches.
This CL ensures of the cleanup in both cases to prevent orphan patch from
both side APM/AF.
If output is still is use (may host direct client(s) or other bridge(s)),
this CL also ensure its patch handle is reused and rerouting is not forced
if no output device change. It will prevent glitches.

Test: (manuel) make a first playback on an SwOutput, stop it, create a SwAudio
Patch, and release it twice. Playback Patch shall always be released,
hense HAL API called.
(functional on emulator):
adb shell ./data/AudioPolicyEmulatorTests --gtest_filter=*AudioPatchTest*
adb shell ./data/AudioPolicyEmulatorTests --gtest_filter=*AudioSourceTest*

Signed-off-by: Francois Gaffie <francois.gaffie@renault.com>
Change-Id: I24e371cd69e04aba2bf74b88336f6323cd88ad8b